git-svn-id: http://svn.resourcespace.com/svn/rs/branches/20160531_astec_imagemagick_optimizations@8595 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/collections_functions.php
@@ -142,11 +142,24 @@ function get_collection_resources($collection)
 	return sql_array("select resource value from collection_resource where collection='$collection' order by sortorder asc, date_added desc, resource desc"); 
 	}
 	
-function add_resource_to_collection($resource,$collection,$smartadd=false,$size="")
+function add_resource_to_collection($resource,$collection,$smartadd=false,$size="",$addtype="")
 	{
-	global $collection_allow_not_approved_share;
-	if (collection_writeable($collection)||$smartadd)
-		{	
+	global $collection_allow_not_approved_share, $collection_block_restypes;	
+	$addpermitted=collection_writeable($collection) || $smartadd;
+	if ($addpermitted &&(count($collection_block_restypes)>0))
+		{
+		if($addtype=="")
+			{
+			$addtype=sql_value("select resource_type value from resource where ref='" . $resource . "'",0);
+			}
+		if(in_array($addtype,$collection_block_restypes))
+			{
+			$addpermitted=false;
+			}
+		}
+		
+	if ($addpermitted)	
+		{
 		# Check if this collection has already been shared externally. If it has, we must fail if not permitted or add a further entry
 		# for this specific resource, and warn the user that this has happened.
 		$keys=get_collection_external_access($collection);
@@ -1300,6 +1313,7 @@ function get_search_title($searchstring){
 
 function add_saved_search_items($collection)
 	{
+	global $collection_share_warning, $collection_allow_not_approved_share, $userref, $collection_block_restypes;
 	# Adds resources from a search to the collection.
 	$results=do_search(getvalescaped("addsearch",""), getvalescaped("restypes",""), "relevance", getvalescaped("archive","",true),-1,'',false,getvalescaped("starsearch",""),false,false,getvalescaped("daylimit",""));
 
@@ -1307,39 +1321,50 @@ function add_saved_search_items($collection)
 	# for this specific resource, and warn the user that this has happened.
 	$keys=get_collection_external_access($collection);
 	$resourcesnotadded=array(); # record the resources that are not added so we can display to the user
-	if (count($keys)>0)
+	$blockedtypes=array();# Record the resource types that are not added 
+	
+	for ($r=0;$r<count($results);$r++)
 		{
-		# Set the flag so a warning appears.
-		global $collection_share_warning, $collection_allow_not_approved_share;
-		$collection_share_warning=true;
+		$resource=$results[$r]["ref"];
+		$archivestatus=$results[$r]["archive"];
 		
-		for ($n=0;$n<count($keys);$n++)
+		if(in_array($results[$r]["resource_type"],$collection_block_restypes))
 			{
-			# Insert a new access key entry for this resource/collection.
-			global $userref;
-			
-			for ($r=0;$r<count($results);$r++)
+			$blockedtypes[] = $results[$r]["resource_type"];
+			continue;
+			}
+
+		if (count($keys)>0)
+			{			
+			if ($archivestatus<0 && !$collection_allow_not_approved_share)
 				{
-				$resource=$results[$r]["ref"];
-				$archivestatus=$results[$r]["archive"];
-				if ($archivestatus<0 && !$collection_allow_not_approved_share) {$resourcesnotadded[$resource] = $results[$r];continue;}
+				$resourcesnotadded[$resource] = $results[$r];
+				continue;
+				}
+			for ($n=0;$n<count($keys);$n++)
+				{
+				# Insert a new access key entry for this resource/collection.
 				sql_query("insert into external_access_keys(resource,access_key,user,collection,date) values ('$resource','" . escape_check($keys[$n]["access_key"]) . "','$userref','$collection',now())");
 				#log this
-				collection_log($collection,"s",$resource, $keys[$n]["access_key"]);
+				collection_log($collection,"s",$resource, $keys[$n]["access_key"]);	
+				
+				# Set the flag so a warning appears.
+				$collection_share_warning=true;	
 				}
 			}
-		}
-
+		}	
+		
+		
 	if (is_array($results))
-		{
+		{		
 		$modifyNotAdded = hook('modifynotaddedsearchitems', '', array($results, $resourcesnotadded));
 		if (is_array($modifyNotAdded))
 			$resourcesnotadded = $modifyNotAdded;
 
 		for ($n=0;$n<count($results);$n++)
 			{
-            $resource=$results[$n]["ref"];
-			if (!isset($resourcesnotadded[$resource]))
+			$resource=$results[$n]["ref"];
+			if (!isset($resourcesnotadded[$resource]) && !in_array($results[$n]["resource_type"],$collection_block_restypes))
 				{
 				sql_query("delete from collection_resource where resource='$resource' and collection='$collection'");
 				sql_query("insert into collection_resource(resource,collection) values ('$resource','$collection')");
@@ -1347,7 +1372,7 @@ function add_saved_search_items($collection)
 			}
 		}
 
-	if (!empty($resourcesnotadded))
+	if (!empty($resourcesnotadded) || count($blockedtypes)>0)
 		{
 		# Translate to titles only for displaying them to the user
 		global $view_title_field;
@@ -1356,7 +1381,16 @@ function add_saved_search_items($collection)
 			{
 			$titles[] = i18n_get_translated($resource['field' . $view_title_field]);
 			}
+		if(count($blockedtypes)>0)
+			{
+			$blocked_restypes=array_unique($blockedtypes);
+			// Return a list of blocked resouce types
+			$titles["blockedtypes"]=$blocked_restypes;
+			}
 		return $titles;
+		}
+	if(count($blockedtypes)>0)
+		{
 		}
 	return array();
 	}
@@ -1640,7 +1674,7 @@ function send_collection_feedback($collection,$comment)
 function copy_collection($copied,$current,$remove_existing=false)
 	{	
 	# Get all data from the collection to copy.
-	$copied_collection=sql_query("select * from collection_resource where collection='$copied'","");
+	$copied_collection=sql_query("select cr.resource, r.resource_type from collection_resource cr join resource r on cr.resource=r.ref where collection='$copied'","");
 	
 	if ($remove_existing)
 		{
@@ -1653,8 +1687,10 @@ function copy_collection($copied,$current,$remove_existing=false)
 	foreach($copied_collection as $col_resource)
 		{
 		# Use correct function so external sharing is honoured.
-		add_resource_to_collection($col_resource['resource'],$current,true);
+		add_resource_to_collection($col_resource['resource'],$current,true,"",$col_resource['resource_type']);
 		}
+	
+	hook('aftercopycollection','',array($copied,$current));
 	}
 
 if (!function_exists("collection_is_research_request")){
@@ -1773,23 +1809,33 @@ function collection_max_access($collection)
 	return $maxaccess;
 	}
 
-function collection_min_access($collection)	
-	{
-	# Returns the minimum access (the least permissive) that the current user has to the resources in $collection.
-	$minaccess=0;
-	if (is_array($collection)){$result=$collection;}
-	else {
-		$result=do_search("!collection" . $collection,"","relevance",0,-1,"desc",false,"",false,"");
-	}
-	for ($n=0;$n<count($result);$n++)
-		{
-		$ref=$result[$n]["ref"];
-		# Load access level
-		$access=get_resource_access($result[$n]);
-		if ($access>$minaccess) {$minaccess=$access;}
-		}
-	return $minaccess;
-	}
+function collection_min_access($collection)
+    {
+    # Returns the minimum access (the least permissive) that the current user has to the resources in $collection.
+    $minaccess = 0;
+    if(is_array($collection))
+        {
+        $result = $collection;
+        }
+    else
+        {
+        $result = do_search("!collection{$collection}", '', 'relevance', 0, -1, 'desc', false, '', false, '');
+        }
+
+    for($n = 0; $n < count($result); $n++)
+        {
+        $ref = $result[$n]['ref'];
+
+        # Load access level
+        $access = get_resource_access($result[$n]);
+        if($access > $minaccess)
+            {
+            $minaccess = $access;
+            }
+        }
+
+    return $minaccess;
+    }
 	
 function collection_set_public($collection)
 	{
@@ -1848,20 +1894,16 @@ function collection_set_themes($collection,$themearr)
 	
 function remove_all_resources_from_collection($ref){
 	// abstracts it out of save_collection()
-		# Remove all resources?
-	if (getval("removeall","")!="")
+	$removed_resources = sql_array('SELECT resource AS value FROM collection_resource WHERE collection = ' . $ref . ';');
+
+	// First log this for each resource (in case it was done by mistake)
+	foreach($removed_resources as $removed_resource_id)
 		{
-		$removed_resources = sql_array('SELECT resource AS value FROM collection_resource WHERE collection = ' . $ref . ';');
-
-		// First log this for each resource (in case it was done by mistake)
-		foreach($removed_resources as $removed_resource_id)
-			{
-			collection_log($ref, 'r', $removed_resource_id, ' - Removed all resources from collection ID ' . $ref);
-			}
-
-		sql_query('DELETE FROM collection_resource WHERE collection = ' . $ref);
-		collection_log($ref, 'R', 0);
+		collection_log($ref, 'r', $removed_resource_id, ' - Removed all resources from collection ID ' . $ref);
 		}
+
+	sql_query('DELETE FROM collection_resource WHERE collection = ' . $ref);
+	collection_log($ref, 'R', 0);
 	}	
 
 if (!function_exists("get_home_page_promoted_collections")){
@@ -2105,7 +2147,7 @@ function compile_collection_actions(array $collection_data, $top_actions, $resou
         }
 
     // Request all
-    if($count_result > 0)
+    if($count_result > 0 && ($k == '' || $internal_share_access))
         {
         # Ability to request a whole collection (only if user has restricted access to any of these resources)
         $min_access = collection_min_access($result);
@@ -2122,7 +2164,7 @@ function compile_collection_actions(array $collection_data, $top_actions, $resou
             $o++;
             }
         }
-	
+
 	if($geo_locate_collection)
         {
             $data_attribute['url'] = sprintf('%spages/geolocate_collection.php?ref=%s',
@@ -2403,3 +2445,30 @@ function compile_collection_actions(array $collection_data, $top_actions, $resou
     return $options;
     }
 }
+
+/**
+* Make a filename unique by appending a dupe-string.
+*
+* @param array $base_values
+* @param string $filename
+* @param string $dupe_string
+* @param string $extension
+* @param int $dupe_increment
+*
+* @return string Unique filename
+*/
+function makeFilenameUnique($base_values, $filename, $dupe_string, $extension, $dupe_increment = null)
+    {
+    // Create filename to check if exist in $base_values
+    $check_filename = $filename . ($dupe_increment ? $dupe_string . $dupe_increment : '') . '.' . $extension;
+
+    if(!in_array($check_filename, $base_values))
+        {
+        // Confirmed filename does not exist yet
+        return $check_filename;
+        }
+
+    // Recursive call this function with incremented value
+    // Doing $dupe_increment = null, ++$dupe_increment results in $dupe_increment = 1
+    return makeFilenameUnique($base_values, $filename, $dupe_string, $extension, ++$dupe_increment);
+    }
